+++ ndk/CHANGELOG.md
@@ -4,6 +4,7 @@
 - media_format: Expose `MediaFormat::copy()` and `MediaFormat::clear()` from API level 29. (#449)
 - input_queue: Add `from_java()` constructor, available since API level 33. (#456)
 - event: Add `from_java()` constructors to `KeyEvent` and `MotionEvent`, available since API level 31. (#456)
+- event: Implement `SourceClass` `bitflag` and provide `Source::class()` getter. (#458)
 
 # 0.8.0 (2023-10-15)